Refactor out most of the species class
