+++ agr_api/src/test/java/org/alliancegenome/api/tests/integration/AbstractIT.java
@@ -1,7 +1,6 @@
 package org.alliancegenome.api.tests.integration;
 
 import org.alliancegenome.core.config.ConfigHelper;
-import org.alliancegenome.neo4j.view.OrthologyModule;
 import org.junit.Before;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -18,7 +17,6 @@ public class AbstractIT {
 
 		mapper.disable(MapperFeature.DEFAULT_VIEW_INCLUSION);
 		mapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
-		mapper.registerModule(new OrthologyModule());
 	}
 
 }
