+++ agr_api/src/test/java/org/alliancegenome/api/tests/integration/PhenotypeIT.java
@@ -25,7 +25,6 @@ import org.alliancegenome.es.model.query.FieldFilter;
 import org.alliancegenome.es.model.query.Pagination;
 import org.alliancegenome.neo4j.entity.EntitySummary;
 import org.alliancegenome.neo4j.entity.node.GeneticEntity;
-import org.alliancegenome.neo4j.view.OrthologyModule;
 import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
@@ -66,7 +65,6 @@ public class PhenotypeIT {
 
 		mapper.disable(MapperFeature.DEFAULT_VIEW_INCLUSION);
 		mapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
-		mapper.registerModule(new OrthologyModule());
 	}
