+++ agr_indexer/src/main/java/org/alliancegenome/indexer/indexers/curation/TransgenicAlleleCurationIndexer.java
@@ -101,7 +101,7 @@ public class TransgenicAlleleCurationIndexer extends Indexer {
 				TransgenicAlleleSummaryDocument document = documentMap.computeIfAbsent(da.getAllele(), allele -> {
 					TransgenicAlleleSummaryDocument doc = new TransgenicAlleleSummaryDocument();
 					doc.setAllele(allele);
-					doc.setPhylogeneticSortingIndex(allele.getTaxon().getPhylogeneticSortOrder());
+					doc.setPhylogeneticSortingIndex(allele.getTaxon().getSpecies().getPhylogeneticOrder());
 					return doc;
 				});
 				List<TransgenicAlleleConstruct> constructList = document.getTransgenicAlleleConstructs();
@@ -126,7 +126,7 @@ public class TransgenicAlleleCurationIndexer extends Indexer {
 				construct.setTargetedGenes(getTargetedGenes(constructObj));
 				document.setHasDiseaseAnnotations(da.getHasDiseaseAnnotations());
 				document.setHasPhenotypeAnnotations(da.getHasPhenotypeAnnotations());
-				document.setPhylogeneticSortingIndex(da.getAllele().getTaxon().getPhylogeneticSortOrder());
+				document.setPhylogeneticSortingIndex(da.getAllele().getTaxon().getSpecies().getPhylogeneticOrder());
 				constructList.add(construct);
 			}
 			display.progressProcess(response.getReturnedRecords());
