--- agr_java_core/src/main/java/org/alliancegenome/neo4j/view/OrthologyModule.java
@@ -1,10 +1,0 @@
-package org.alliancegenome.neo4j.view;
-
-import com.fasterxml.jackson.databind.module.SimpleModule;
-
-public class OrthologyModule extends SimpleModule {
-	public OrthologyModule() {
-		super("OrthologyModule");
-	}
-
-}
